#25: feat: add structure files
