+++ tests/.env.example
@@ -1,5 +1,9 @@
 REDIS_HOST=redis
 REDIS_PORT=6379
+REDIS_USER=default
+REDIS_PASSWORD=redis
+REDIS_DB_INDEX=0
+REDIS_DSN=redis://${REDIS_USER}:${REDIS_PASSWORD}@${REDIS_HOST}:${REDIS_PORT}/${REDIS_DB_INDEX}
 
 
 ELASTICSEARCH_DSN=http://elasticsearch:9200/
