+++ tests/functional/src/test_search.py
@@ -1,6 +1,6 @@
 import pytest
 
-from functional.testdata.search_param import SEARCH_PARAM
+from tests.functional.testdata.test_param.search_param import SEARCH_PARAM
 
 @pytest.mark.parametrize(
     'query_data, expected_answer, index, endpoint, generate, map',
@@ -18,11 +18,26 @@ async def test_films_search(
     status, body_count, timestamp1 = await make_get_request(
         endpoint, query_data
     )
-    assert status == expected_answer['status']
-    assert body_count == expected_answer['length']
+    assert status == expected_answer['status'], (
+        f'Проверьте, что get-запрос к `/api/v1/{endpoint}` '
+        f'возвращает ответ с кодом {expected_answer['status']}.'
+    )
+    assert body_count == expected_answer['length'], (
+        f'Проверьте, что get-запрос к `/api/v1/{endpoint}` '
+        f'возвращает в ответе нужное количество {expected_answer['length']}.'
+    )
     status, body_count, timestamp2 = await make_get_request(
         endpoint, query_data
     )
-    assert status == expected_answer['status']
-    assert body_count == expected_answer['length']
-    assert timestamp1 > timestamp2
+    assert status == expected_answer['status'], (
+        f'Проверьте, что get-запрос к `/api/v1/{endpoint}` '
+        f'возвращает ответ с кодом {expected_answer['status']}.'
+    )
+    assert body_count == expected_answer['length'], (
+        f'Проверьте, что get-запрос к `/api/v1/{endpoint}` '
+        f'возвращает в ответе нужное количество {expected_answer['length']}.'
+    )
+    assert timestamp1 > timestamp2, (
+        f'Проверьте, что get-запрос к `/api/v1/{endpoint}` '
+        f'возвращает ответ из кеш Redis.'
+    )
